Remove dead LocationWindow components, close NaN known issue
lib/components/location.tsx (the parseFloat-of-empty-string NaN source) is
imported nowhere; the live location path (useNoteLocation, api_calls, map
screen) already guards against NaN. The time.tsx date picker component was
equally dead, so time.tsx shrinks to formatToLocalDateString and becomes .ts,
and @react-native-community/datetimepicker is dropped. Tests for the dead
components removed; formatToLocalDateString gets direct coverage.
