+++ package.json
@@ -19,7 +19,6 @@
     "@expo/metro-runtime": "~56.0.16",
     "@expo/vector-icons": "^15.1.1",
     "@react-native-async-storage/async-storage": "2.2.0",
-    "@react-native-community/datetimepicker": "9.1.0",
     "@react-native-community/slider": "5.2.0",
     "@tanstack/react-query": "^5.100.9",
     "babel-plugin-transform-inline-environment-variables": "^0.4.4",
